# GH-45624: [C++][Parquet] BitReader optimize case for num_bits == 1