### Rationale for this change

The benchmark shows that, in `parquet-column-reader-benchmark`, the `GetValue_` takes half of the times. We can try to optimize this.

### What changes are included in this PR?

Avoid call `detail::GetValue_` if possible.

### Are these changes tested?

Covered by existing

### Are there any user-facing changes?

No


* GitHub Issue: #45624